@@ -1,6 +1,6 @@
 {
   "name": "blockheads-api",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "description": "APIs for interacting with cloud servers and mac servers without requiring changes to your code.",
   "main": "cloud.js",
   "types": "could.d.ts",
@@ -9,20 +9,19 @@
   "license": "GPL-3.0",
   "dependencies": {
     "simple-plist": "^0.2.1",
-    "util.promisify": "^1.0.0"
+    "typed-promisify": "^0.4.0"
   },
   "devDependencies": {
-    "@types/tape": "^4.2.30",
-    "tap-spec": "^4.1.1",
-    "tape": "^4.7.0",
-    "ts-node": "^3.1.0",
-    "tslint": "^5.4.3",
-    "typedoc": "^0.7.1",
-    "typescript": "^2.4.1"
+    "ava": "^0.22.0",
+    "nyc": "^11.2.1",
+    "tslint": "^5.7.0",
+    "typedoc": "^0.7.2",
+    "typescript": "^2.5.3"
   },
   "scripts": {
     "build": "tsc",
-    "test": "tape -r ts-node/register **/*.test.ts | tap-spec",
+    "test": "tsc --module CommonJS && nyc ava",
+    "report": "nyc report --reporter=html --reporter=lcov",
     "docs": "typedoc --out docs --mode modules --name Blockheads-Api --excludeNotExported --excludePrivate && echo Save >> docs/.nojekyll",
     "prepare": "npm run build"
   }
